refactor(design-library): extract generic MarkdownMessage component
Move the MarkdownMessage component from apps/web into
packages/design-library with a pluggable linkComponent prop,
decoupling the generic markdown rendering from Vellum-specific
OAuth popup link handling.

The app-layer wrapper at apps/web/src/components/markdown-message.tsx
is now a thin re-export that injects the OAuthAwareLink component,
keeping all existing consumer imports unchanged.

Changes:
- Add MarkdownMessage, MarkdownMessageProps, MarkdownLinkComponent
  exports to @vellum/design-library
- Add react-markdown@10.1.0 and remark-gfm@4.0.1 as design-library
  dependencies
- Add data-slot="markdown-message" to root element
- Split tests: generic rendering tests in design-library, OAuth
  behaviour tests in apps/web
- Exclude .test.tsx from design-library tsconfig (matches existing
  .test.ts exclusion)
