+++ packages/design-library/package.json
@@ -50,6 +50,8 @@
     "class-variance-authority": "0.7.1",
     "clsx": "2.1.1",
     "lucide-react": "1.16.0",
+    "react-markdown": "10.1.0",
+    "remark-gfm": "4.0.1",
     "sonner": "2.0.7",
     "tailwind-merge": "3.6.0"
   }
